+++ firmware/hdl/packages/gt_mp7_core_pkg.vhd
@@ -1,14 +1,15 @@
 
 -- actual versions:
 -- use "FRAME_VERSION" as mp7_ugt release fw version (used for tag name).
--- mp7_ugt (=FRAME_VERSION): v1.21.0
+-- mp7_ugt (=FRAME_VERSION): v1.22.0
 
 -- use "GT_VERSION" as mp7_ugt release fw version (used for tag name).
--- gt: v1.21.0
--- frame: v1.3.0 (see frame.vhd)
+-- gt: v1.22.0
+-- frame: v1.4.0 (see frame.vhd)
 -- gtl: v1.17.2 (see gtl_module_tpl.vhd)
 -- fdl: v1.4.0 (see fdl_module.vhd)
 
+-- HB 2022-09-20: v1.22.0 - Changed top_decl.vhd to 8 REGIONs. Changed output_mux.vhd and lmp.vhd (frame: v1.4.0).
 -- HB 2022-05-31: v1.21.0 - Signal start_sync_bc0_int used for reset of prescale counter (instead of begin_lumi_section). Removed sres signals for counters in FDL, not used anymore.
 -- HB 2021-05-31: v1.20.2 - Updated script run_simulation_questa.py (removed UGT_GITLAB_PWD)
 -- HB 2021-04-25: v1.20.1 - Updated and added scripts.
@@ -75,7 +76,7 @@ package gt_mp7_core_pkg is
 -- ==================================================================================================
 -- GT firmware version
     constant GT_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant GT_MINOR_VERSION      : integer range 0 to 255 := 21;
+    constant GT_MINOR_VERSION      : integer range 0 to 255 := 22;
     constant GT_REV_VERSION        : integer range 0 to 255 := 0;
 	constant GT_VERSION : std_logic_vector(31 downto 0) := X"00" &
            std_logic_vector(to_unsigned(GT_MAJOR_VERSION, 8)) &
@@ -83,7 +84,7 @@ package gt_mp7_core_pkg is
            std_logic_vector(to_unsigned(GT_REV_VERSION, 8));
 -- FRAME firmware version
     constant FRAME_MAJOR_VERSION      : integer range 0 to 255 := 1;
-    constant FRAME_MINOR_VERSION      : integer range 0 to 255 := 3;
+    constant FRAME_MINOR_VERSION      : integer range 0 to 255 := 4;
     constant FRAME_REV_VERSION        : integer range 0 to 255 := 0;
 -- GTL firmware (fix part) version
     constant GTL_FW_MAJOR_VERSION      : integer range 0 to 255 := 1;