+++ firmware/hdl/packages/top_decl.vhd
@@ -29,15 +29,18 @@ package top_decl is
         constant ZS_ENABLED: boolean := FALSE;
         constant CLOCK_AUX_RATIO: clock_ratio_array_t := (1, 1, 1);
 
--- HB 2016-11-24: N_REGION = 7 (no more needed for ugt in- and outputs - see mp7_brd_decl.vhd)
+-- HB 2022-09-20: N_REGION = 8 (no more needed for ugt in- and outputs - see mp7_brd_decl.vhd)
+-- Reserved REGION 4 for 5g ZDC optical links
+-- Moved read-out-record outputs to REGION 5..7 (links 20..29)
         constant REGION_CONF: region_conf_array_t := (
                 0 => (gth_10g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_10g_std_lat, 3, 10), -- 0 / 118
                 1 => (gth_10g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_10g_std_lat, 3, 10), -- 1 / 117*
                 2 => (gth_10g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_10g_std_lat, 3, 10), -- 2 / 116
                 3 => (gth_10g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_10g_std_lat, 4, 11), -- 3 / 115
-                4 => (gth_10g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_10g_std_lat, 4, 11), -- 4 / 114*
+                4 => (gth_5g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_5g_std_lat, 4, 11), -- 4 / 114*
                 5 => (gth_10g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_10g_std_lat, 4, 11), -- 5 / 113
-                6 => (gth_10g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_10g_std_lat, 5, 12) -- 6 / 112
+                6 => (gth_10g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_10g_std_lat, 5, 12), -- 6 / 112
+                7 => (gth_10g_std_lat, u_crc32, buf, no_fmt, buf, u_crc32, gth_10g_std_lat, 5, 12) -- 7 / 111*
         );
 
 end top_decl;