+++ firmware/hdl/payload/frame/output_mux.vhd
@@ -54,6 +54,8 @@ end output_mux;
 
 architecture arch of output_mux is
 
+    constant begin_output_links: positive := 20; 
+
     signal sValid : std_logic := '0';
 
     signal readout_finor : std_logic_vector(31 downto 0);
@@ -100,7 +102,7 @@ begin
             in3     =>  s_in3_mux0,   -- frame 3   -> algo_after_gtLogic 96-127
             in4     =>  s_in4_mux0,   -- frame 4   -> algo_after_gtLogic 128-159
             in5     =>  s_in5_mux0,   -- frame 5   -> algo_after_gtLogic 160-191
-            mux_out =>  lane_out(16)
+            mux_out =>  lane_out(begin_output_links+1)
         );
 
 
@@ -123,7 +125,7 @@ begin
             in3     =>  s_in3_mux1,    -- frame 3   -> algo_after_gtLogic 288-319
             in4     =>  s_in4_mux1,    -- frame 4   -> algo_after_gtLogic 320-351
             in5     =>  s_in5_mux1,    -- frame 5   -> algo_after_gtLogic 352-383
-            mux_out =>  lane_out(17)
+            mux_out =>  lane_out(begin_output_links+2)
         );
 
 
@@ -147,7 +149,7 @@ begin
             in4     =>  s_in4_mux2,    -- frame 4   -> free
             in5     =>  s_in5_mux2,    -- frame 5   -> free
             -- sel     =>  frame_cntr,
-            mux_out =>  lane_out(18)
+            mux_out =>  lane_out(begin_output_links+3)
         );
 
 
@@ -170,7 +172,7 @@ begin
             in3     =>  s_in3_mux3,   -- frame 3   -> algo_after_bxomask 96-127
             in4     =>  s_in4_mux3,   -- frame 4   -> algo_after_bxomask 128-159
             in5     =>  s_in5_mux3,   -- frame 5   -> algo_after_bxomask 160-191
-            mux_out =>  lane_out(19)
+            mux_out =>  lane_out(begin_output_links+4)
         );
 
 
@@ -193,7 +195,7 @@ begin
             in3     =>  s_in3_mux4,    -- frame 3   -> algo_after_bxomask 288-319
             in4     =>  s_in4_mux4,    -- frame 4   -> algo_after_bxomask 320-351
             in5     =>  s_in5_mux4,    -- frame 5   -> algo_after_bxomask 352-383
-            mux_out =>  lane_out(20)
+            mux_out =>  lane_out(begin_output_links+5)
         );
 
 
@@ -217,7 +219,7 @@ begin
             in4     =>  s_in4_mux5,    -- frame 4   -> free
             in5     =>  s_in5_mux5,    -- frame 5   -> free
             -- sel     =>  frame_cntr,
-            mux_out =>  lane_out(21)
+            mux_out =>  lane_out(begin_output_links+6)
         );
 
 
@@ -240,7 +242,7 @@ begin
             in3     =>  s_in3_mux6,   -- frame 3   -> algo_after_prescaler_mask 96-127
             in4     =>  s_in4_mux6,   -- frame 4   -> algo_after_prescaler_mask 128-159
             in5     =>  s_in5_mux6,   -- frame 5   -> algo_after_prescaler_mask 160-191
-            mux_out =>  lane_out(22)
+            mux_out =>  lane_out(begin_output_links+7)
         );
 
 
@@ -263,7 +265,7 @@ begin
             in3     =>  s_in3_mux7,    -- frame 3   -> algo_after_prescaler_mask 288-319
             in4     =>  s_in4_mux7,    -- frame 4   -> algo_after_prescaler_mask 320-351
             in5     =>  s_in5_mux7,    -- frame 5   -> algo_after_prescaler_mask 352-383
-            mux_out =>  lane_out(23)
+            mux_out =>  lane_out(begin_output_links+8)
         );
 
 
@@ -287,7 +289,7 @@ begin
             in3     =>  s_in3_mux8,    -- frame 3   -> algo_after_prescaler_mask 480-511
             in4     =>  s_in4_mux8,    -- frame 4   -> finor
             in5     =>  s_in5_mux8,    -- frame 5   -> free
-            mux_out =>  lane_out(24)
+            mux_out =>  lane_out(begin_output_links+9)
         );
 
     -- JW 2015-08-24: added local and mp7 bc_cntr to output
@@ -322,7 +324,7 @@ begin
             in3     =>  s_in3_mux9,    -- frame 3   -> free / orbit counter 47..32
             in4     =>  s_in4_mux9,    -- frame 4   -> free / orbit counter 31..0
             in5     =>  s_in5_mux9,    -- frame 5   -> free
-            mux_out =>  lane_out(25)
+            mux_out =>  lane_out(begin_output_links+10)
         );
 
 end architecture;