+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v1.22.0] - 2022-09-20
+### Comment
+
+- mp7_ugt firmware release v1.22.0
+- changes for ZDC optical connections
+
+### Changed
+- source files:
+  - ../payload/frame/output_mux.vhd
+  - ../payload/frame/lmp.vhd
+  - ../packages/top_decl.vhd
+  - ../packages/gt_mp7_core_pkg.vhd
+
 ## [v1.21.0] - 2022-09-14
 ### Comment